refactor: 직사각형과 원의 중복 코드 제거
